+++ src/router/PrivateRoute.jsx
@@ -1,0 +1,11 @@
+import { Navigate } from 'react-router-dom';
+import { useSelector } from 'react-redux';
+import { selectIsLoggedIn } from '../store/authSlice';
+
+function PrivateRoute({ children }) {
+  const isLoggedIn = useSelector(selectIsLoggedIn);
+
+  return isLoggedIn ? children : <Navigate to="/login" replace />;
+}
+
+export default PrivateRoute;